+++ rest_framework/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '2.2.6'
+__version__ = '2.2.7'
 
 VERSION = __version__  # synonym